Разгрузка Statistics
- Выделен псевдо-базовый класс для графиков
- Вынесена логика формирование BarChart в сам класс (как пример альтернативной реализации)
- Вынесены из методов инициализации коллекций
- Небольшой рефакторинг
